price int->long

+++ app/src/main/java/com/example/usager/projectfinalandroid/ActivityMarket.java
@@ -113,19 +113,19 @@ public class ActivityMarket extends AppCompatActivity {
 
         Button btnM=(Button) findViewById(R.id.btnManual);
         Button btnA =(Button) findViewById(R.id.btnAutomatic);
-        int prixM=Integer.valueOf(btnM.getText().toString()) ;
-        int prixA=Integer.valueOf(btnA.getText().toString()) ;
+        long prixM=Integer.valueOf(btnM.getText().toString()) ;
+        long prixA=Integer.valueOf(btnA.getText().toString()) ;
         //prixM=30*Joueur.getManuel()+100;
         //prixA=30*Joueur.getAutomatic()+100;
-        prixM=(int)formuleScalePrice(15,1.15,Joueur.getManuel())+100;
-        prixA=(int)formuleScalePrice(15,1.15,Joueur.getAutomatic())+100;
+        prixM=(long)formuleScalePrice(15,1.15,Joueur.getManuel())+100;
+        prixA=(long)formuleScalePrice(15,1.15,Joueur.getAutomatic())+100;
         btnM.setText(String.valueOf(prixM));
         btnA.setText(String.valueOf(prixA));
 
     }
     public long formuleScalePrice(int a, double r, int n)
     {
-        long ex=(int)java.lang.Math.pow(r,(int)n);
+        long ex=(long)java.lang.Math.pow(r,(long)n);
         double price=a*(ex - 1)/(r - 1);
         return (long) price;
     }
